Separated scheduler from AIDO class for better readability
