+++ example_comparison.py
@@ -1,5 +1,5 @@
 from container_examples.calo_opt.interface_calo_opt import AIDOUserInterfaceExample
-from modules.aido import AIDO
+from modules.scheduler import AIDO
 
 if __name__ == "__main__":
     AIDO.optimize([
