+++ sim_param_dict
@@ -1,0 +1,1 @@
+{"absorber_thickness": {"name": "absorber_thickness", "starting_value": 10.0, "optimizable": true, "sigma": 1.0, "units": null, "min_value": 0.5, "max_value": 40.0, "current_value": 10.0, "discrete_values": null, "cost": null}, "absorber_material": {"name": "absorber_material", "starting_value": "LEAD", "optimizable": true, "sigma": null, "units": null, "min_value": null, "max_value": null, "current_value": "LEAD", "discrete_values": ["LEAD", "TUNGSTEN"], "probabilities": [0.5, 0.5], "cost": null}, "energy": {"name": "energy", "starting_value": 1000, "optimizable": false, "sigma": null, "units": null, "min_value": null, "max_value": null, "current_value": 1000, "discrete_values": null, "cost": null}, "num_absorber_plates": {"name": "num_absorber_plates", "starting_value": 2, "optimizable": true, "sigma": null, "units": null, "min_value": null, "max_value": null, "current_value": 2, "discrete_values": [0, 1, 2, 3, 4], "probabilities": [0.2, 0.2, 0.2, 0.2, 0.2], "cost": null}, "metadata": {"iteration": 0, "creation_time": "2024-10-23 17:55:23.400767", "description": ""}}
